gles2: Make new targets current immediately
Currently the Gles2Renderer sets new targets on `Bind`-implementations,
but does not setup the GL-state to actually use them, because it expects
the next command to call `make_current` anyway. This is correct for calls
to `Gles2Renderer`, but not `Gles2Frame`. So if downstream would change
the target inside a frame (e.g. by using `copy_texture`), like the
multigpu renderer does we end up in a bad state.