Fix surface-damage indefinitely accumulating
wayland/compositor is designed to expect downstream to drain/clear
the accumulated damage from `SurfaceData` on commit. To handle
multiple renderers the renderer/utils module did not do that, causing
damage to pile up indefinitely. This is obviously wrong, so this commit
attempts to fix this, by accumulating damage into `SurfaceState`
(clearing `SurfaceData` in the process) and keeping at most four commits
of damage for different renderers to consume.