desktop: do not round the frame output size and... ...ceil the output geometry size.
this fixes the frame output size by not scaling and rounding
at all which would introduce rounding issues.
the ceil makes sure that at least the output size is
cleared/drawn which fixes flickering at the edge of an output.
the flickering is caused by a mismatch between clear and drawing
a window that extends over the edge.