fix xdg destroy protocol checks
moves the xdg destroy checks from the
destoryed handler to the destroy request
handler. this is done to ensure sequential processing
of the destroy events as the destroyed handler
are executed delayed and do not guarantee
ordering. ordering is important as the protocol
specifies that a xdg role object has to be destroyed
before destroying the xdg surface