vulkan: remove Vulkan1X features/properties from PhysicalDevice
The Vulkan1X types are not ideal for our level of abstraction nor too idomatic. Smithay intends to support higher versions of Vulkan like a superset of existing functionality. However the Vulkan1X types cannot be chained with their individual parts without validation errors. There is also the fact that these types are not available until Vulkan 1.2, and the Vulkan11 structures are actually provided in Vulkan 1.2, not Vulkan 1.1. Therefore for the sake of cleaning up the codebase, the Vulkan1X types now need to be obtained some other way if desired.