android: load the x11 assets as a tarball to the cache directory
This require the x11 assets from xkeyboard-config to be loaded as a tarball in `assets/x11.tar` of the APK.

I'm not sure why, but it seems AssetManager has an issue if we're dealing with complex filetrees like the X11 assets folder!
So we can pack them as a tarball and unpack them to fix it!

Original: 62e2229a7123d983f817276fd0c5fd55d73f56ac

Signed-off-by: Nefo Fortressia <nefothingy@hotmail.com>